+++ crates/libwebrtc-sys/include/adm.h
@@ -116,5 +116,5 @@ class OpenALPlayoutADM : public webrtc::AudioDeviceModuleImpl {
   ALCcontext* _playoutContext = nullptr;
   ALCdevice* _playoutDevice = nullptr;
 
-  std::mutex _mutex;
+  std::recursive_mutex _mutex;
 };